+++ src/main/java/by/toukach/restservlet/service/PhoneNumberService.java
@@ -12,9 +12,9 @@ public interface PhoneNumberService {
 
     void update(PhoneNumberDTO phoneNumberDTO) throws NotFoundException;
 
-    PhoneNumberDTO findById(Long phoneNumberId) throws NotFoundException;
+    PhoneNumberDTO findById(int phoneNumberId) throws NotFoundException;
 
     List<PhoneNumberDTO> findAll();
 
-    boolean delete(Long phoneNumberId);
+    boolean delete(int phoneNumberId);
 }
